Replace make with meson
Using meson simplifies the build, and may make it easier to use this
code in project using meson like qemu.

This change uses meson to build blkhash library, and use the library in
the blksum and blkhash_test executables.

Running the tests with meson is not great, since it does not show the
test output, but the complete test output is available in the build
directory.

Since meson support multiple build directories, running the tests
requires passing pytest the path to the blksum executable, which makes
it harder to run the tests manually. On the other hand, the now we can
test installed blksum which may be useful.

Fixes: #4