blksum: Compute blkhash per segment
Preparing for parallel computation, change checksum to:

    H( H(segment 1) + H(segment 2) ... H(segment N) )

When H(segment N) is:

    H( H(block 1) + H(block 2) ... H(block M) )

Segment size is 128 MiB. Depending on image size, the last segment may
be shorter. If the image is not aligned to block size (64 KiB), the last
block of the last segment may be shorted than block size.

With this scheme, we can compute the checksum for multiple segments in
parallel when using file or nbd source.

When using pipe no concurrency is possible, but the checksum is created
in the same way so we can compare checksums created via pipe to checksum
created via other sources.