blksum: Resolve digest earlier
Check if digest_name exists early and fail with a better error message.

Keep digest and digest_size as globals, since we used them later for
creating the root hash.

Since we use same the same digest everywhere, and we already know the
digest size, we don't need to pass pointers around to get the size.