blksum: Parallel checksum
When using file or nbd source, use multiple threads to speed up the
checksum computation.

We use up to 4 worker threads, each processing one segment, and copying
the segment digest to the digest buffer. When all threads finish, the
digest buffer contains all the segment digests in the right order, and
we compute the root hash from the digests.

Fixes #13