parallel: Keep initial source in job
Instead of keeping the original filename in job, keep the initial source
and use it to access the size. The source is kept open while the job is
active. The workers use the job source uri to open their source instead
of the original filename.

This will allow opening qemu-nbd for local file and returning nbd
source, to be cloned by the workers.

Keeping the initial source open during the job will allow using nbd
systemd socket activation to run qemu-nbd. The cost of open file
descriptor or inactive nbd handle is very low.