Use qemu-nbd for local file
When using a filename, run qemu-nbd and use nbd source. This fixes the
checksum for qcow2 images, and use less resources for sparse raw images.

This is a huge usability improvement, but it may be slower for raw
images on fast local storage. We may want to use qemu-nbd only if file
is mostly empty.