Update benchmarks
Replace the benchmarks using NBD URL with simpler filename, since we use
NBD internally anyway. Testing show that running NBD server manually and
using NBD URL is just as fast as starting qemu-nbd and stopping it for
real images.

The benchmarks show that processing raw images is slower now for images
with lot of data on local storage. I think the best way to optimize raw
images is implementing extents in the file source, and using qemu-nbd
only for qcow2 images.