Keep image format when opening nbd server
Keep image format in the src struct. It is set to the image format when
opening an image using nbd server, and to "raw" when opening a file,
since file source treat any image as raw format. Format is always NULL
when using NBD URL, since NBD server do not report the image format.