Probe once when building without NBD
open_src() accepts now a format argument. This allows the caller to
avoid probing filename again when opening a new source for the same
file.

When initializing a job, we open the source with format=NULL to probe
filename. When initializing a worker, we use the job source format to
avoid additional probing.