Adaptive queue size
Testing shows that for image with large extents, queue-depth of 4 is
best, but for image with tiny extents queue-depth of 16-64 is best.
Instead of limiting the number of items in the queue, limit the total
number of bytes of all in-flight commands.

Zero commands do not start nbd async commands, but we don't want to
queue too many zero commands, so we consider zero command as a tiny
read.

The --queue-depth option was renamed to --queue-size, and the default
value is now 1 MiB. With the default --read-size, we queue 4-64
commands.

Comparing to previous version shows up to 10% speedup and up to 20%
reduction in system time.

$ hyperfine -w1 -r3 -L v fixed,adaptive "{v}/blksum sha1 -n fedora-34.raw"

Benchmark 1: fixed/blksum sha1 -n fedora-34.raw
  Time (mean ± σ):     10.238 s ±  0.247 s    [User: 30.650 s, System: 13.465 s]
  Range (min … max):    9.980 s … 10.472 s    3 runs

Benchmark 2: adaptive/blksum sha1 -n fedora-34.raw
  Time (mean ± σ):      9.361 s ±  0.118 s    [User: 29.766 s, System: 10.795 s]
  Range (min … max):    9.231 s …  9.462 s    3 runs

Summary
  'adaptive/blksum sha1 -n fedora-34.raw' ran
    1.09 ± 0.03 times faster than 'fixed/blksum sha1 -n fedora-34.raw'

$ hyperfine -w1 -r3 -L v fixed,adaptive "{v}/blksum sha1 -n 1m-extents.raw"

Benchmark 1: fixed/blksum sha1 -n 1m-extents.raw
  Time (mean ± σ):      9.651 s ±  0.040 s    [User: 36.835 s, System: 6.653 s]
  Range (min … max):    9.617 s …  9.696 s    3 runs

Benchmark 2: adaptive/blksum sha1 -n 1m-extents.raw
  Time (mean ± σ):      9.405 s ±  0.108 s    [User: 36.585 s, System: 5.297 s]
  Range (min … max):    9.338 s …  9.530 s    3 runs

Summary
  'adaptive/blksum sha1 -n 1m-extents.raw' ran
    1.03 ± 0.01 times faster than 'fixed/blksum sha1 -n 1m-extents.raw'