Simplify process_segment()
Move the logic to consume data from the current extent to next_extent().
This function returns now the next extent to process, consuming all or
some data from the current extent, and advancing to the next extent when
current extent is consumed.

Change has_extent() to consider the bytes left in the current extent. If
has_extent() returns true, we can call next_extent() to consume the next
extent.

Simplify functions that received zero and len arguments to accept an
extent.