Handle short extents reply
If the nbd server returns a short reply, we used to skip part of the
segment silently, creating incorrect checksum. This can happen if a
segment has more than 131072 extents, since qemu-nbd limit the reply
size to 1 MiB (NBD_MAX_BLOCK_STATUS_EXTENTS).

This is only a theoretical issue, since when using qcow2 format, the
minimum size of an extent is 64 KiB, so we can have up to 2048 extents
per segment, and for raw image, since file systems allocate 4 KiB for
single byte, we can have up to 32768 extents per segment. But it is more
robust to have correct code.

Change the extents handling loop to end when offset reaches the end of
the segment. This happens when we start the last command for this
segment.

In each iteration, we fetch more extents from the source if needed, and
get the next extent to process. This ensures that after a short reply we
will fetch more extents from the nbd server.