Limit number of queued commands
The adaptive limit allows too many tiny commands even in the default value, and
when using larger read size and queue size, the limit is not effective.

Limit the number of commands to 16 per worker, same as qemu-nbd per connection
limit.

With this change, we can use:

    --queue-size 4m --read-size 2m

Which works best when using --cache, without queuing 256 4k commands.

Testing show identical performance for local and iscsi images, and 3-5% speedup
and 5-10% lower cpu usage with nfs.