Improve command queue naming
Replace "len" with "bytes" to make the purpose of it more clear. Replace
"count" with "len", which is the common term for number of items.

Since now "len" includes now zero commands, we can simplify the cost() function
and return 0 for zero command. This allows sending up to 16 in-flight nbd
requests.

For example, if we use:

    --queue-size 4m --read-size 2m

We can queue now two 2m read commands, even if there are zero commands in the
queue.