More flexible test image format string
The previous format could describe images bigger than one segment, so
the test could not verify correct segment processing.

Change the format to pairs of size:byte separated by whitespace. With
this can specify an image of any practical size.

Examples:

2 GiB empty image:

    create_image(path, "1g:-")

768 MiB sparse image with 64k block every 128 MiB:

    create_image(path, "131008k:- 64k:A 131008k:- 64k:B 131008k:- 64k:C")

Computing simple checksum for large images is slow, so we are not going
to create big images in the tests, but we can create image with multiple
segments.