Split tests with fixtures
Previously for every image we run multiple tests using single test
invocation. If nbd was not supported part of the test were skipped
silently. This makes it hard to understand which tests ran, and which
variant failed.

Change the tests to create the test images with fixtures.
- image - create a raw image for every format string
- raw - create checksum for every raw image for every tested digest
- qcow2 - create qcow2 image for every raw fixture

The fixture use session scope, so we create an image an compute a
checksum once.

Split the tests to separate test per variant:
- raw file
- qcow2 file
- raw pipe
- raw nbd
- qcow2 nbd

The tests depending on nbd are skipped properly, so missing nbd
dependencies are clear.