parallel: Exit on read error
Previously if nbd server returned an error, we would silently ignore the
error in the completion callback, and update the hash with invalid data
return from malloc() when the command buffer was allocated.

Now we exit with an error message with the failed error code. It would
be nicer to get an error message from libnbd, but I'm not sure we can
call nbd_get_error() from a completion callback, and if it returns
meaningful value during a callback.