Avoid pointless progress updates
A progress bar can show only 0-100 percents, so is no point in
rendering the progress bar again if already shows the same value.

Keep the last value and skip the update if the value did not change.
This limits the number of updated to 100, regardless of the size of the
image.

Testing shows 7.5% speedup when using large terminal window, for example
window maximized to full 2k screen, but no difference with the default
80x24 window.

Large window:

    $ time blksum sha1 -p fedora-34.raw
    e3bd64cad6f27beac93115858d85ca9b06949682  fedora-34.raw

    real    0m10.036s
    user    0m23.295s
    sys     0m14.314s

    $ time build/blksum sha1 -p fedora-34.raw
    e3bd64cad6f27beac93115858d85ca9b06949682  fedora-34.raw

    real    0m9.336s
    user    0m22.327s
    sys     0m13.370s

Small window:

    $ time blksum sha1 -p fedora-34.raw
    e3bd64cad6f27beac93115858d85ca9b06949682  fedora-34.raw

    real    0m8.642s
    user    0m22.076s
    sys     0m11.981s

    $ time build/blksum sha1 -p fedora-34.raw
    e3bd64cad6f27beac93115858d85ca9b06949682  fedora-34.raw

    real    0m8.633s
    user    0m22.067s
    sys     0m11.983s

Fixes #35.