Suppress qemu-nbd stderr
qemu-nbd stderr is not helpful since blksum get error info as a client,
and in some cases it is way to verbose, logging an error for every
failed nbd request.

Unless we run in debug mode (BLKSUM_DEBUG=1), qemu-nbd stderr is
redirected now to /dev/null.

To be able to report execvep() errors, we save the original stderr and
restore it after errors.

If we cannot open /dev/null for writing, stderr is not redirected, and
we log a warning.