Unify the way blksum is run in the tests
Add Blksum wrapper for running blksum and use it in all tests.

Waiting for the wrapper communicates with the child process and collect
the results of the run so we can assert about them.

For the termination tests, the wrapper supports waiting until qemu-nbd
socket is created and finding qemu-nbd pid.