Drain extra blocks after stopping
When the worker thread start handling errors, it may fail when
processing a block. The rest of the blocks in the queue will leak.

This can also happen in normal condition if blkhash submit more blocks
after the special zero-length final block. This should not happen, but
if it does, the blocks would leak.

Fix by draining the queue after a worker was stopped, and waking the
producer thread if it was blocked on a full queue.

More work is needed to ensure that no new blocks are submitted after
a worker stops.