Handle errors in worker_update()
If the worker failed, updating a worker fails now with the error that
failed the worker. Submitting blocks to the worker may leak the blocks.

The pthread functions should never fail in normal conditions, but if
they fail it can lead to reporting bad checksum or to a deadlock. The
function tries to release the lock and report the error to the caller.

A caller should stop updating the worker, finalize it, and fail the
operation.

If worker_update() fails, the caller is responsible for deallocating the
block. Document this in the header. Current code just leaks the block in
this case.