Handle errors in pop_block()
If pthread_mutex_* or pthread_cond_* fail (very unlikely), set the
error, stop the worker, and signal the error by returning a NULL block.

The caller will discover the failure on the next call to worker_update()
or worker_digest() and should stop all workers and fail the entire
operation.