Handle errors in EVP_Digest* functions
Since we allocate the context when creating the worker, the EVP_Digest*
should never fail, but it they do we cannot ignore the error. Now we
check all calls, set the error and stop the worker.

When calling the functions modifying the hash, return early if the
worker already failed. This makes the worker thread main loop simpler.