Handle errors when stopping workers
When we stopped the workers we did not handle errors from
`workers_final()` and `workers_digest()`. This could lead to missing an
error in the worker and computing a wrong checksum.

Fixed by checking all errors and returning an error if finalizing or
getting a digest failed.

Checking error is delicate since we must call `worker_final()` and
`worker_digest()` on all workers. The first call will stop the worker if
it is running, and the second call will join the thread and compute the
final digest for the worker.

Once we detected a failure, we don't add the failed worker digest to the
root hash.

The code for stopping workers was extracted to `stop_workers()` helper.
`blkhash_free()` is using it to stop the workers without computing the
final hash instead of calling `blkhash_final()` and dropping the result.