Handle errors when updating workers
When consuming pending data or consuming user data, allocating a block
or updating the worker can fail.

- If allocating a block failed, we ignored the error and tried to
  update the worker with NULL block, which will segfault (good).

- If worker_update() failed, we ignored the error. This leaks the block
  (bad), and may lead to deadlock when the worker queue becomes full
  (very bad).

- Since `consume_data()` did not return any error, `consume_pending()`
  could not report the error to `blkhash_final()`. This could lead to
  reporting a wrong checksum.

All errors are checked now, and the internal function report them so
`blkhash_update()` and `blkhash_zero()` can report the error in the next
step.

If `blkhash_final()` detect an error in `consume_pending()` it stops the
workers quickly without computing a checksum and report the error to the
caller.

Part of #57