Handle blkhash errors in blksum simple source
If blkhash fails we print an error, free the hash, and exit with an
error.

Previously we could try to exit without calling blkhash_free() which can
lead to segfault during exit when openssl atexit handlers run while
worker threads are using opessnl functions. I think this issue is the
reason for random failures in the CI.

Part of #57