Handle blkhash errors in blksum parallel
If `blkhash_update()` or `blkhash_zero()` we need to break the read
loop and call `blkhash_free()` so the worker threads are terminate
before the main thread exit.

This is not possible to do with the current code. The current code calls
`FAIL()` for any error which logs the failure message, marks the process
as failed, and exit the worker thread. In this case `blkhash_free()` is
not called so it is possible that the worker threads fill segfault when
openssl run the atexit handlers when the main thread exit.

We need to improve this later, it may the reason for random CI failures.

Part of #57