Support non-final zero length blocks
Previously if the worker receives a zero length block it treated the
block as the last block and terminated after handling the block. Add a
`last` flag to block to mark the final block.

With this change we can send zero length blocks during normal operation,
signaling the worker that they need to add zero block since the last
block up to the zero length block.