Eliminate segments
Splitting image to segments is not needed now. Instead of waiting until
all commands finish every 128m, we can process the entire image, and get
extents from the source when needed.

This will allow increasing extent size and minimizing the number of
extents calls if a server reports extents after the requested range.