Don't clip extents requests
NBD server can return a long reply - an extent ending after the
requested range. This was unwanted when we processed the image in 128m
segments, but now this can be very useful, minimizing the number of
extents calls and speeding up processing of large unallocated areas.

Testing with qemu-nbd shows that it does not return long replies:

$ BLKSUM_DEBUG=1 build/blksum /data/tmp/empty-1t.raw 2>&1 | grep 'get extents' | head
[  0.006505] worker 0 get extents offset=0 length=134217728
[  0.006621] worker 0 get extents offset=134217728 length=134217728
[  0.006658] worker 0 get extents offset=268435456 length=134217728
[  0.006693] worker 0 get extents offset=402653184 length=134217728
[  0.006724] worker 0 get extents offset=536870912 length=134217728
[  0.006753] worker 0 get extents offset=671088640 length=134217728
[  0.006780] worker 0 get extents offset=805306368 length=134217728
[  0.006810] worker 0 get extents offset=939524096 length=134217728
[  0.006837] worker 0 get extents offset=1073741824 length=134217728
[  0.006866] worker 0 get extents offset=1207959552 length=134217728

With 1t empty image it could return 4g zero extent on every request.

Fixes #66