Split project to sub directories
Clean up the mess, separating the project to sub directories:

- include - include files consumed by most components
- lib - the blkhash library, used by blksum and the tests
- common - common sources used by tests and blksum
- bin - the blksum command
- test - tests and benchmarks
- doc - the man pages and example sources

Every directory has a small and simple meson.build file. The main
meson.build file also became small and simple.

I hope this will make it easier to use blkhash as meson subproject, but
even if not, this seems easier to understand and maintain.

Part-of: #60