Fix unity unwanted installation
The reason for the unwanted unity installation is unity PR[1] enabling
installation of the static library, pkgconfig file, and headers.

We got this commit because we use `head` instead of a latest release.
Fix by pinning unity to v2.5.2, which does not have the issue, and
remove the unneeded --skip-subprojects option.

[1] https://github.com/ThrowTheSwitch/Unity/pull/658