Add option to build without the manual pages
On macOS building the manual pages is broken now. It seems that both
xmllint and xsltproc from macports have issues.

Add a `doc` meson option to allow building without the manual pages. To
use it setup the build directory with:

    meson setup build -Ddoc=disabled

NOTE: This will break the rpm spec on Linux.

Since this option is a hack for macOS it is not documented.