Add threads arguments in blkhash_new()
This is problematic now since changing the number of threads change the
hash value, but it helps to test scalability.

Since this changes the public interface, this is good time to change
also the argument order so the first argument is the digest name. We
should have a better way to add new options without breaking
compatibility but at this point this is still ok.

Part-of: #79