Fix possible double-free on errors
If a block was added to the queue in worker_update() but signaling the
worker thread or unlocking the lock failed, the function was failing and
the caller was freeing the block. Later the block could be free by the
worker thread.

The requirement to free the block on errors was also confusing and
required a comment.

Simplify by always consuming the block. If the block was not added to
the queue worker_update() free the block and return an error.

Part-of: #79