Simplify stopping workers
Previously we sent the total size of the image to worker_final() and it
submitted the last block with the end index to all workers. The worker
was filling in zero blocks if needed until the end of the image.

Now we send explicit zero block at end index before stopping the
workers, so worker_final() does not need the size. The image_size is not
used now so it was removed.

Part-of: #79