Separate stream from worker
Preparing for having different numbers of workers and streams, extract
the code computing block hashes and the stream hash to a new module.

We keep now 2 lists - workers and streams. When submitting blocks we
first locate the stream handling the block using:

    stream_index = block_index % config.streams

Then we find the worker for this stream using:

    worker_index = stream_index % config.workers

Since the current code always create the same number of streams and
workers, nothing was changed yet, and changing the number of workers
will change the hash.

This also greatly simplify the worker code. Now we have 2 simpler
modules, each doing simpler task.

Part-of: #79