Configurable number of streams in blkhash
With this change we can configure also the number of streams (e.g. 32).
Since the hash depends on the number of streams, we get the same hash
value for any number of threads, allowing user configurable threads.

Making streams user configurable is problematic, since changing the
number of streams changes the hash value. But it may be required for
some applications so I'm keeping this option available for the library.

This change keeps the default number of streams, keeping the hash
compatible with previous version.

Part-of: #79