Future proof the API
Improve ease of use and interoperability by making it easier to create a
hash with default options:

    h = blkhash_new();

This creates a hash that will be compatible with blksum or with other
users of the blkhash library using the default constructor.

User that want to use non-default options can create a hash that will
not work with other users hashes can create a blkhash_opt struct and use
the new constructor accepting it:

    opts = blkhash_opts_new("blake2b512");
    if (opts == NULL)
        return errno;

    err = blkhash_opts_set_block_size(opts, 256 * 1024);
    if (err)
        return err;

    err = blkhash_opts_set_threads(opts, 16);
    if (err)
        return err;

    h = blkhash_new_opts(opts);
    blkhash_opts_free(opts);

This is very verbose but it lets your control everything. This is also
what we do in the tests, benchmarks, and in blksum, because we want to
test all the options.

Like the struct blkhash, struct blkhash_opt is opaque and must be
allocated by the library, so future version of the library can add new
fields without breaking existing users.

The user can access the struct only via the accessors, so we can
validate the options when the user sets them, and fail early. When we
use a blkhash_opts input, we know that it must be valid since there is
no way to create invalid one.