Increase default streams to 32
Change the default streams value to 32 to allow up to 32 threads without
changing the hash value. Testing show that even on big machines using
more than 16 threads is not useful in most case. Use the next power of 2
to allow for future improvement without changing the hash value.

The default number of streams is exposed via blkhash.h to make to keep
the blkhash library, blksum tool, and 3rd party tools using the library
compatible.

The test were modified to use 32 streams and test 1, 2, 4, 8, 16, 32
threads. Since changing the streams value changes the hash, we had to
update the expected values.

Part-of: #79