Add benchmark timeout
We worked hard to tune the input size so the benchmark run for at least
one second so we report stable results, but not too much time so we get
fast results. Tuning the input size is also endless pain, requires
changing each time we improve the code.

Fix by adding --timeout-seconds argument, set to 1 second by default, so
the benchmark does the right thing out of the box. This greatly
simplifies blkhash-bench.py, and useful for quick demos.

If --input-size is specified, it overrides the timeout and the benchmark
computes a checksum for the specified size. This is useful for
validating the correctness of the benchmark program, used by
bench_test.py.