Add submission completion
A submission can reference now a completion event, allowing the thread
calling blkhash_update() to synchronize with the worker threads.

If a submission is created with a non-null completion, it increases the
completion reference count. When a submission is freed, it decreases the
completion reference count. When the last reference is removed, the
completion callback is invoked.

This will become a public API when we introduce the async update.