Add threads helpers
After initializing a mutex or condition, failure to use them is unlikely
and we don't have a way to recover, so the best way to handle the error
is to abort with a message, so we have a core dump to inspect.

The current code tried to propagate the error which makes the code too
complicated and hard to maintain, and in case of error, impossible to
debug since we return the error cleanly.

Add helpers functions to call pthread functions that must not fail and
abort on errors.