Avoid implicit digest fetching
We used EVP_get_digestbyname() which is a compatibility layer for using
openssl < 3. When using this function, the returned digest does not
contain the implementation, so it is fetched implicitly when using
EVP_DigestInit_ex(). The docs recommend to use explicit fetching for
best performance.

Rename look_digest() to create_digest() and add free_digest() helper to
free the returned value. The helpers hide the differences between
openssl 1.1 (RHEL 8) and 3 (Fedora 37).

[1] https://www.openssl.org/docs/man3.1/man7/crypto.html#Explicit-fetching