Benchmark the async API
Add --aio option to benchmark the async API. In this mode --queue-depth
controls the number of in-flight requests.  The defaults are based on
blksum defaults.

Split the benchmark to simple zero_size() and zero_timeout() for testing
blkhash_zero(), and more complicated update_size() and update_timeout()
for testing blkhash_update() or blkhash_aio_update().