Report error in completions
We would like to report errors, so completion keeps an error.  The error
will be set by the first failing worker and reported in the completion
callback.

The completion is allocated now in completion_new(), and freed when the
last reference is removed in completion_unref(). This requires
additional allocation per blkhash async update, but profiling does not
show that malloc have interesting overhead. We can consider using
pre-allocated completions later if needed.