Detect zero in workers in async mode
If we need to copy submission data it is more efficient to detect zeros
in the update thread, but otherwise the work is better done in the
worker threads, speeding work dispatching to the worker threads, and
getting higher throughput since zero detection can be done in parallel.

Testing shows 40% slowdown for single worker (unlikely scenario), but at
least 50% speedup with 2 threads or more. With 4 threads we see 384%
speedup (98 GiB/s). I don't think we will see any difference with real
images since the bottleneck is reading data from storage.