Limit threads by number of online cpus
Due to the way blkhash distribute work to threads, using 16 threads on a
12 cpus machine can be faster, but after that typically the results are
worse. Make bench.threads() smarter by limiting the number of threads to
the first power of 2 bigger or equal to the number of online cpus.

This will save time testing on smaller machines (for example in the CI
or on a build machine), and avoid wasting time on improving the results
for uninteresting cases (e.g. 32 threads on 12 cores machine).