Extract create_hash helper
Creating the hash is complicated and it will be even more complicated
when we create event fds. Also there is no reason to create the hash and
free it in the worker thread. Move the code to new helper called from
init_worker(), and free the hash in destroy_worker().