Simplify queue size management
Previously we had a special --queue-size that tried to limit the number
of bytes in in-flight requests. The motivation was to get better
performance on laptops when using smaller queue size (2 MiB) give
slightly better result. This mechanism is not very effective and add
complexity that make it hard to use the new async API.

Replace the --queue-size option with simpler --queue-depth setting the
size of the queue. Use the new option to limit the number of in flight
requests.