Eliminate command_queue
This was added to simplify the code managing the queue limiting the
number of items by total size of all requests. Now that we have simpler
way, this struct only make it harder. Inline it in the worker struct.

The length of the queue was renamed to commands_in_flight. This will be
useful later to make the read/update loop simpler.

The queue was renamed to read_queue since it is used only for reading.
When we start using the async API we will add a new queue for updates.