Use the async API in blksum
Split run_update() to start_update() and update_completed(), since now
it behaves like start_read(). Since start_read() is async and run_zero()
is not, updating bytes_hashed and progress moved to update_completed()
and run_zero().

wait_for_event() polls now both blkhash completion fd and source fd. We
don't need to skip the NBD fd, since we always have the blkhash event
fd, and we call this function only if we have commands in flight.

When we get completion events from blkhash, we read the data from the
completion fd, get the completions and complete the in flight updates.