Add hash queue
Instead of allocating a command for every read and deallocating when
update or zero complete, pre-allocated all commands during startup.

Commands are kept in 2 queues:
- read queue - single link tail queue, keeping commands in order
- hash queue - double link tail queue, keeping commands while they are
  hashed.

read_more_data() move completed commands from the hash queue to the read
queue and start new reads. hash_more_data() move the completed commands
from the had of the read queue to the hash queue, and run update or
zero.