Improve commands_in_flight logic
We increase the count when we start an aync command, and decrease
when the async command completes. The count is not use now for limiting
the number of commands in the queue since the commands are
pre-allocated.

The counter is not use for anything now except for validating that we
respect the queue depth, but it can be used later for waiting for
events and for debugging.