Signal completion only when needed
Signal completions only when moving from empty completion queue to
non-empty. Otherwise we already sent a signal so the caller should wake
up and get the completion without sending a signal.

This change complicates getting completions. If the user does not
consume all completions (no good reason to do this), they will not get a
new notification since the queue is not empty. In this case we signal
the completion fd so on the next poll the user will get the
notification.