Support up to 64 threads
With the async API we can process up to 10.6 GiB/s with 64 threads.
Increase the default streams value to 64 to allow 1-64 threads without
changing the hash value.

This change breaks compatibility with older versions, changing the hash
value.

The examples in `test/README.md` use now the async API and 64 threads to
show the new capabilities.