Simplify src_extents
Instead of replicating the nbd interface allocating extents array on
every call, accept extents array owned by the caller, and copy the
extents from the server to the caller array.

If the server returned too many extents, they are dropped and the caller
will fetch them on the next call. This way we eliminate allocations in
this code path and simplify aio-checksum.

Maximum extent is set to 4096, using 32k. This is seems much more than
needed for all the images I tested.