Fix x axis scale
Add the results as integer value to get the right scale on the x axis.
This is best done in tow passes, first converting the results to list of
x, y, kwargs tuples, and then plotting them.

This way we can inspect the shape of the results when setting xticks.
Current we assume that the first result contains all ticks but this can
be improve later.