Test handling of many extents
Add tests for images with many extents (8192 extents in 512 MiB image).
For qcow2 image this is one extent for every cluster (64K).

Running with DEBUG logs shows that we handle extents correctly - fill up
the extents array with 4096 extents and drop the rest, and ask for more
extents in the next request:

    $ BLKSUM_DEBUG=1 build/bin/blksum /var/tmp/blkhash/extents0/raw 2>&1 | grep extents
    [  0.005079] Get extents offset=0 length=536870912
    [  0.010522] Got 4096 extents in 5440 usec
    [  0.131145] Get extents offset=268435456 length=268435456
    [  0.133851] Got 4096 extents in 2705 usec