Simplify gettime()
On macOS we can use clock_gettime() with CLOCK_UPTIME_RAW. The result is
the same as calling mach_absolute_time() and converting the value to
nanoseconds using timebase info.

On Linux we should use CLOCK_MONOTONIC_RAW which is not affected by NTP
adjustments.

On FreeBSD we should use CLOCK_UPTIME for the same semantics.

Otherwise we default to CLOCK_MONOTONIC which hopefully does the right
ting on other platforms.