Improve docs/blkhash-performance.md
Describe the ways used to improve performance and include the most
important graphs from the test results directory, and some details on
how we test blkhash.

Add a link to the blkhash performance page from the readme, and improve
formatting of the links section.

Add a scaled images to the media directory for displaying in the
blkhash-performance page.

Fixes #94