Start digest abstraction layer
The digest layer provides a uniform way to create, init, update,
finalize and free a digest using any digest from openssl or other
libraries (e.g. blake3, xxhash).

This change adds the digest module and implement the "null" digest. This
is same as openssl "null" digest, but it is useful to implement it for
testing the infrastructure and for simplifying the openssl backend.