Include openssl digests in digest_list()
This is a much simpler version of bin/list_digest.c, moving memory
management to the caller, providing array to copy into the digests.
The caller can get the number of digests by calling with NULL array.

Listing digests should be exposed as a public blkhash API and blksum
should use it instead of accessing openssl.