Add blkhash_digests() public API
Users like blksum need this to get the list of available digests.

Move the implementation to the digest module since it is small enough
and having everything related to digests in one module makes it easier
to work with the code.

Move the test to blkhash-test since this is now a public API.