Add blake3-devel to containers and CONTRIBUTING.md
To build with blake3 support we need the blake3-devel[1] package. It is
available for Fedora 38,39,40, and for RHEL/Centos via EPEL 8,9.

For macOS we need the bake3 brew package and a little bit of
configuration to find the libraries and headers.

This changes modifies only for the fedora containers. We will handle the
centos containers later.

[1] https://packages.fedoraproject.org/pkgs/blake3/blake3-devel/