Experiment with mmap
It is very easy to map the entire image, and feed blkhash with pieces of
memory from the map instead of allocating buffers and reading data from
the image.

The actual reading happens in the kernel when the blkhash worker thread
access the memory for zero detection or computing the block hash. This
is the same model used by b3sum.