Experiment again with a flat merkle tree
The first idea when we worked on decoupling the number of threads from
the hash construction was using a flat merkele tree:

    H( H(block 1) || H(block 2) ... H(block N) )

We rejected this because building the root hash cannot be done in
parallel, trying to optimize for huge unallocated images. I think this
was a mistake, and we need to optimize for the allocated areas of an
image, which takes most of the time.

A flat merkle tree has several advantages:
- We can compute the block hashes using a simple thread pool, using any
  number of threads instead of power of 2.
- A simple thread pool can work better when we have different cores
  (e.g.  performance cores, efficiency cores).
- A simple thread pool works better when some blocks take less time to
  compute (zero block). If a thread finishes early, it can handle the
  next block in the queue.
- Uses less resources (1 root hash instead of 64)