Allow changing submission type
Preparing for handling submissions in a thread pool, allow changing a
DATA submission to ZERO. This can be used when a thread finds that a
submission data is a zero block. The thread can change the submission
type to zero so when we add the md to the root hash we will add a zero
block md.

With the current code this is not needed since the same thread add the
zero block md to the hash immediately.