Implement flat merkle tree
Replace the streams based construction with flat merkle tree:

    H( H(block 1) || H(block 2) || ... || H(block N) )

To maintain block order, submissions are are kept in a submission FIFO
queue managed by the blkhash thread.

Zero blocks are queued in the submission queue, and added to the root
when they reach the front of the queue.

Data blocks are queued in the submission queue, and submitted to a
thread pool for detecting zeros and computing the block hash.  When the
operation is completed, the hash are added in order to the root hash.

When the thread pool detects a zero blocks, it marks the submission as
zero block and complete the submission. On the submission queue, the
blkhash thread add a zero block hash to the root hash.

If submitting to the thread pool failed, or computing a block hash
failed, the submission is marked as failure. The blkhash thread will
fail the entire operation when detecting a failed submission.

Unfinished:
- remove streams option from blkhash_opts
- update manual and docs