Fix image caching in blksum benchmarks
When running without enough memory to cache all the test images at once,
sometimes images are not in the page cache during the entire run, and
the results show wrong results.

Fixed by adding a helper to cache and uncache an image, and using it
around the hyperfine command.

Tested only on Linux, may require more work for BSD or macOS.